fix: reverted changes in contact list, because we still need the dummy contact data
